myntra_presets/python_api/requirements.txt
fastapi[standard]
uvicorn
requests
jinja2
pyrebase4
python-dotenv